+++ compiler-core/src/ast/typed.rs
@@ -1220,6 +1220,20 @@ impl TypedExpr {
         }
     }
 
+    /// If this expression refers to a record constructor, returns the name of
+    /// that constructor. This is the unqualified name for `Var` and the label
+    /// for a qualified `ModuleSelect`.
+    ///
+    pub fn record_constructor_name(&self) -> Option<EcoString> {
+        if let TypedExpr::Var { name, .. } = self {
+            Some(name.clone())
+        } else if let TypedExpr::ModuleSelect { label, .. } = self {
+            Some(label.clone())
+        } else {
+            None
+        }
+    }
+
     /// If the given expression is a literal record builder, this will return
     /// index of the variant being built.
     ///